# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Owoore API
  version: 1.0.0
  description: |
    Multi-tenant digital giving platform for churches — dedicated virtual bank
    accounts per member per fund, automatic reconciliation, and M-of-N
    signatory payout governance. Built on Nomba.

    ![Two of Owoore's core calls — registering a church and verifying a member OTP](/docs/owoore-hero.svg)

    ## Get started
    There's no shared public API key here — Owoore is multi-tenant, so every
    church gets its own admin account and its own token from a live call, not
    a static sandbox credential pasted into docs.

    1. **Register a church** — `POST /api/v1/orgs` with a name, admin name,
       email, and password. This one call also creates the default payout
       policy and the default Tithe/Offering funds.
    2. **Verify the admin email** — an OTP is sent on registration; confirm
       it with `POST /api/v1/auth/admin/verify-email` to activate the account
       and receive your first admin JWT.
    3. **Authenticate every admin request after that** with
       `Authorization: Bearer <token>` from step 2, or from
       `POST /api/v1/auth/admin/login` on subsequent visits.
    4. **Members join separately** — `POST /api/v1/auth/send-otp` then
       `POST /api/v1/auth/verify-otp` against the church's join slug, no
       admin token involved.

    > **Note:** every endpoint below is grouped by resource in the sidebar —
...

    ## Money conventions (read this first)
...

...

...

...
servers:
...
components:
  securitySchemes:
    adminToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin JWT (roles ADMIN / TREASURER / SIGNATORY)
    memberToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Member JWT from OTP verification
  schemas:
...
    Error:
      type: object
      properties:
        success: { type: boolean, const: false }
        error:
          type: object
          properties:
            code: { type: string, example: VALIDATION_ERROR }
            message: { type: string }
...
  responses:
...
    Forbidden:
      description: Wrong role or wrong organisation (tenant isolation)
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: Resource not found (always org-scoped)
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
paths:
...
  /api/v1/orgs/{slug}:
    get:
      tags: [Organisations]
      summary: Public church lookup by join slug
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: '`{ id, name, slug, logo_url, joinLink }`' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/orgs/{id}:
    patch:
      tags: [Organisations]
      summary: Update church details (own org only)
      security: [{ adminToken: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Updated org }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/v1/webhooks/nomba:
    post:
      tags: [Webhooks]
      summary: Nomba event sink (HMAC-verified, durable, idempotent)
      description: |
        **Signature**: HMAC-SHA256 over the colon-joined concatenation of
        `event_type:requestId:merchant.userId:merchant.walletId:
        transaction.transactionId:transaction.type:transaction.time:
        responseCode:timestamp`, keyed with the webhook secret,
        base64-encoded — sent as `nomba-signature` with `nomba-timestamp`.
        The raw request bytes must reach the verifier untouched (the route
        uses a raw body parser).

        **Durability contract**: 200 is returned only after the event is a
        pg-boss row in Postgres. Processing is asynchronous; failures retry
        5× with backoff, then dead-letter (operator alerted).

        **Handled events**: `payment_success` (credit member/shared ledger),
        `payout_success` (settle transfer, debit ledger), `payout_failed` /
        `payout_refund` / `payment_reversal` (release soft lock, mark FAILED),
        `payment_failed` (logged, no action).
      parameters:
        - { name: nomba-signature, in: header, required: true, schema: { type: string } }
        - { name: nomba-timestamp, in: header, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                requestId: { type: string, description: Idempotency key }
                event_type: { type: string }
                data: { type: object }
      responses:
        '200': { description: Event durably queued (or duplicate — also 200) }
        '401': { description: Bad/missing signature or timestamp }

  # ── Funds ────────────────────────────────────────────────────────────
...